edit board
