+++ src/components/Sidebar/BoardList/BoardList.jsx
@@ -1,15 +1,18 @@
 import { BoardListItem } from "../BoardListItem/BoardListItem"
+import { useSelector } from "react-redux"
+import { selectAllBoards } from "redux/boards/selectors"
 
-export const BoardList = ({boards}) => {
-return (
-
+export const BoardList = () => {
+  //const boards = useSelector(selectAllBoards)
+  const boards = [{id: '12415', title: 'ala', icon: 'icon-colors', background: '4'}, {id: '32985', title: 'Board', icon: 'icon-container', background: '13'}]
+  
+  return (
     <ul>
-<li><BoardListItem/></li>
-    {/* {boards.map(board => (
+    {boards.map(board => (
       <li key={board.id}>
         <BoardListItem board={board} />
       </li>
-    ))} */}
+    ))}
   </ul>
-)
+  )
 }
